update the controllers

+++ app/Http/Controllers/Api/TaskController.php
@@ -1,32 +1,57 @@
 <?php
 namespace App\Http\Controllers\Api;
+
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
-use App\Http\Resources\TaskCollection;
 use App\Http\Resources\TaskResource;
+use App\Models\Task;
 
 class TaskController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+    // Admin -> see all tasks
+    // Manager -> see all tasks
+    // Member -> see only assigned tasks
     public function index()
     {
-        $tasks = auth()->user()->tasks()->latest()->get();
-        return $this->success(new TaskCollection($tasks), 'Task list retrieved successfully.');
+        // get the current authenticated user
+        $user = auth()->user();
+        // get the user'role
+        $role = $user->role->title;
+
+        if ($role === 'admin' || $role === 'manager') {
+            return Task::with(['assignee', 'creator'])->get();
+        }
+
+        return Task::with(['assignee', 'creator'])->where('user_id', $user->id)->get();
     }
 
     /**
      * Store a newly created resource in storage.
      */
+    // Admin -> can assign tasks
+    // Manager -> can assign tasks
+    // Member -> cannot assign tasks
     public function store(StoreTaskRequest $request)
     {
+        $user = auth()->user();
+        $role = $user->role->title;
 
-        $task = auth()->user()->tasks()->create([
+        if (! in_array($role, ['admin', 'manager'])) {
+            return response()->json([
+                'message' => 'Unauthorized',
+            ], 403);
+        }
+
+        $task = Task::create([
              ...$request->validated(),
-            'status' => $request->status ?? 'pending',
+            'status'     => $request->status ?? "pending",
+            'created_by' => $user->id,
         ]);
+
         return $this->success(new TaskResource($task), 'A task has been successfully created.');
     }
 
@@ -35,17 +60,34 @@ class TaskController extends Controller
      */
     public function show($id)
     {
-        //$task = Task::where('id', $id)->where('user_id', auth()->id())->firstOrFail();
         $task = auth()->user()->tasks()->findOrFail($id);
         return $this->success(new TaskResource($task), 'Task retrieved successfully.');
     }
 
+    // Admin -> can update anything
+    // Manager -> can update anything
+    // Member -> can update the assigned task (status of the task)
     /**
      * Update the specified resource in storage.
      */
     public function update(UpdateTaskRequest $request, $id)
     {
-        $task = auth()->user()->tasks()->findOrFail($id);
+        $user = auth()->user();
+        $role = $user->role->title;
+        $task = Task::find($id);
+
+        if (! $task) {
+            return response()->json([
+                'message' => 'Task not found',
+            ], 404);
+        }
+
+        if (! in_array($role, ['admin', 'manager'])) {
+            return response()->json([
+                'message' => 'Unauthorized',
+            ], 403);
+        }
+
         $task->update($request->validated());
         return $this->success(new TaskResource($task), 'Task updated successfully');
     }
@@ -68,9 +110,21 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
+    // Admin -> can delet task
+    // Manager -> cannot delet task
+    // Member -> cannot delet task
     public function destroy($id)
     {
-        $task = auth()->user()->tasks()->findOrFail($id);
+        $user = auth()->user();
+        $task = Task::findOrFail($id);
+
+        $role = $user->role->title;
+        if ($role !== 'admin') {
+            return response()->json([
+                "message" => "Unauthorized",
+            ], 403);
+        }
+
         $task->delete();
         return $this->success(null, 'Task deleted successfully.', 200);
     }
